Final 2.0
Delete: Added double click to delete selected song.

Load and Save: Some controls were in public scope fixed them to private,
 ixed pre setted to PlayList now it should be pre setted to Library;
replaced filechooser calling with own class which handles it;
Added a method which adds to queue and Library, if a sobg is added to the Library the ID will be checked, adding to queue has no specific controll structures.


IDGenerator: added method which checks if a specific ID already exists.